Pin r-base (CONDA_R) to major.minor version instead of full version.
This has been discussed in conda-forge/r-base-feedstock#16. The benefit is that
we don't have to rebuild everything for every released patch version.
R does not explicitly say that it follows semantic versioning, however the codebase
gives strong evidence for this:

1. In the sources (base/R/library.R), they have an error if a package was build before 3.0.0, and a warning if a package was built with a newer version than the current. No warning in case of an older version.
2. The function update.packages() from r-base, which is recommended to be executed after updating R and copying over packages to the new version (as e.g. done by installr), has an option checkBuilt, which, when set to true rebuilds packages if and only if the package has been built under an older major.minor version. Not under a changed patch version.

In summary, all built-in facilities are designed to take care for major.minor updates, implying that major.minor.patch updates are safe and need no rebuild.